+++ camp44/models/__init__.py
@@ -2,3 +2,4 @@ from .base import *
 from .user import User, UserCreate, UserRead, UserUpdate  # noqa: F401
 from .app import App, AppCreate  # noqa: F401
 from .entity import Entity, EntityCreate, EntityRead, EntityUpdate  # noqa: F401
+from .demo_booking import DemoBooking, DemoAvailabilityConfig  # noqa: F401